Extract time_input, date_input, text_input, and text_area
